feat(ui): implement public proxy route for Strapi requests with authorization handling
